Reorganize Future class and interface
