+++ src/Workflow.php
@@ -13,7 +13,7 @@ namespace Temporal\Client;
 
 use React\Promise\PromiseInterface;
 use Temporal\Client\Activity\ActivityOptions;
-use Temporal\Client\Future\FutureInterface;
+use Temporal\Client\Transport\FutureInterface;
 use Temporal\Client\Workflow\ActivityProxy;
 use Temporal\Client\Workflow\WorkflowEnvironmentInterface;
 use Temporal\Client\Workflow\WorkflowInfo;
