+++ src/Workflow/WorkflowEnvironment.php
@@ -15,9 +15,9 @@ use JetBrains\PhpStorm\ExpectedValues;
 use React\Promise\CancellablePromiseInterface;
 use React\Promise\PromiseInterface;
 use Temporal\Client\Activity\ActivityOptions;
-use Temporal\Client\Future\Future;
-use Temporal\Client\Future\FutureInterface;
 use Temporal\Client\Support\DateInterval;
+use Temporal\Client\Transport\Future;
+use Temporal\Client\Transport\FutureInterface;
 use Temporal\Client\Transport\Protocol\Command\RequestInterface;
 use Temporal\Client\Worker\Worker;
 use Temporal\Client\Workflow\Command\CompleteWorkflow;
